Allow wallfinding debug without full wallfinding
Previously, the wallfinding debug visualization required
the mesh to be oriented (i.e. for orientationfinder to have run).
This meant that it would only work if a roommodel had been loaded,
which only happens if the wallfinder had successfully completed.

This commit adds an option to wallfindapp to return immediately
after orientationfinding with the orientation transform.