Set confidence in overexposed pixels to nonzero
Pixels with value of 255 were originally given 0 confidence, which
meant that upon reprojection, vertices would effectively get
black values. This fix gives them a very low weight, which should
not significantly affect vertices that are not overexposed in all
images but will improve the behavior of vertices that only have
overexposed pixels.