Compute light amounts explicitly
We overloaded the meaning of the lightamount when introducing
lighting basis functions: previously it was simply a total form factor
(for diffuse lights) from the hemicube, but with lighting bases it also
incorporates the basis coefficients.

We thus separated the form factor term out as an aggregate. This
primarily enabled a more effective way of dealing with unobserved
pixels